Correção do forms de cadastro
