+++ app/api/routes.py
@@ -1,5 +1,6 @@
 from flask import Blueprint, render_template, request, redirect, url_for, session, jsonify
-from app.models import db, Estudo, get_dashboard_stats, EDP, listar_estudos, obter_estudo, Municipio, TipoSolicitacao
+from app.models import db, Estudo, get_dashboard_stats, EDP, listar_estudos, obter_estudo, Municipio, TipoSolicitacao, \
+    Regional, Circuito, RespRegiao, Usuario, Subestacao
 
 api_bp = Blueprint("api", __name__)
 
@@ -13,29 +14,29 @@ def teste():
         } for e in edp]
     )
 
-@api_bp.route('/api/municipios/<int:municipio_id>')
-def api_obter_municipio(municipio_id):
-    try:
-        # Método otimizado que carrega tudo de uma vez
-        municipio = Municipio.query.get_or_404(municipio_id)
-
-        if not municipio:
-            return jsonify({'error': 'Municipio não encontrado'}), 404
-
-        municipio_data = {
-            'id': municipio.id_municipio,
-            'nome': municipio.municipio,
-            'edp': municipio.id_edp
-
-        }
-
-        return jsonify(municipio_data)
-
-    except Exception as e:
-        return {
-            'error': 'Erro ao obter municipio',
-            'message': str(e)
-        }, 500
+# @api_bp.route('/api/municipios/<int:municipio_id>')
+# def api_obter_municipio(municipio_id):
+#     try:
+#         # Método otimizado que carrega tudo de uma vez
+#         municipio = Municipio.query.get_or_404(municipio_id)
+#
+#         if not municipio:
+#             return jsonify({'error': 'Municipio não encontrado'}), 404
+#
+#         municipio_data = {
+#             'id': municipio.id_municipio,
+#             'nome': municipio.municipio,
+#             'edp': municipio.id_edp
+#
+#         }
+#
+#         return jsonify(municipio_data)
+#
+#     except Exception as e:
+#         return {
+#             'error': 'Erro ao obter municipio',
+#             'message': str(e)
+#         }, 500
 
 
 
@@ -109,3 +110,34 @@ def internal_error(error):
 @api_bp.errorhandler(404)
 def not_found(error):
     return jsonify({'error': 'Endpoint não encontrado'}), 404
+
+
+# Rotas AJAX para filtros dinâmicos no forms Cadastrar
+@api_bp.route("/api/municipios/<int:id_edp>")
+def get_municipios_by_edp(id_edp):
+    """API para buscar municípios por EDP"""
+    municipios = Municipio.query.filter_by(id_edp=id_edp).all()
+    return {'municipios': [{'id': m.id_municipio, 'nome': m.municipio} for m in municipios]}
+
+
+@api_bp.route("/api/regionais/<int:id_edp>")
+def get_regionais_by_edp(id_edp):
+    """API para buscar regionais por EDP"""
+    regionais = Regional.query.filter_by(id_edp=id_edp).all()
+    return {'regionais': [{'id': r.id_regional, 'nome': r.regional} for r in regionais]}
+
+
+@api_bp.route("/api/circuitos/<int:id_edp>")
+def get_circuitos_by_edp(id_edp):
+    """API para buscar circuitos por EDP"""
+    circuitos = Circuito.query.filter_by(id_edp=id_edp).join(Subestacao).all()
+    return {'circuitos': [{'id': c.id_circuito, 'nome': f"{c.circuito} - {c.subestacao.nome}"}
+                          for c in circuitos]}
+
+
+@api_bp.route("/api/resp_regioes/<int:id_regional>")
+def get_resp_by_regional(id_regional):
+    """API para buscar responsáveis por regional"""
+    responsaveis = RespRegiao.query.filter_by(id_regional=id_regional).join(Usuario).all()
+    return {'responsaveis': [{'id': r.id_resp_regiao, 'nome': r.usuario.nome}
+                             for r in responsaveis]}
